requirements.txt
# Core agent framework
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-anthropic>=0.3.0
langchain-openai>=0.2.0
langsmith>=0.1.0
langchain-community>=0.3.0

# LLM providers
anthropic>=0.40.0
openai>=1.50.0

# PDF extraction
pymupdf>=1.24.0

# PostgreSQL + pgvector
psycopg2-binary>=2.9.0
pgvector>=0.3.0

# Cloud storage
boto3>=1.35.0

# Fuzzy matching
rapidfuzz>=3.9.0

# Date parsing
python-dateutil>=2.9.0

# Web search
tavily-python>=0.3.0

# Environment
python-dotenv>=1.0.0

# HTTP (fx_tool)
requests>=2.32.0
